Refactor neural network interface and models (#24)
Previously functions in interface.py made sure that the neural net
predictions were formatted correctly for models to use, and then
models would call the correct function. interface.py now directly
exposes the predictions of neural nets via `neural_net_predict`, and
models take in this prediction and format the prediction so that they
can use it.

This commit also removes "model 1", since it was too bad for serious
use.